backend/requirement.txt
fastapi
uvicorn[standard]
psycopg2-binary
asyncpg
pydantic
python-dotenv
sqlglot
ollama
python-dotenv
psycopg2-binary
google-generativeai
